+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTAdditiveExpression.java
@@ -31,8 +31,11 @@ package net.sourceforge.pmd.lang.java.ast;
  *     <tr><td>{@code 1 + 2 - 3 - 4 + 5}</td><td>{@code (((1 + 2) - 3 - 4) + 5)}</td></tr>
  * </table>
  *
+ *
+ * @deprecated Replaced with {@link ASTInfixExpression}
  */
-public final class ASTAdditiveExpression extends AbstractLrBinaryExpr {
+@Deprecated
+public final class ASTAdditiveExpression extends AbstractJavaExpr {
 
     ASTAdditiveExpression(int id) {
         super(id);