+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTMultiplicativeExpression.java
@@ -28,8 +28,11 @@ package net.sourceforge.pmd.lang.java.ast;
  *     <tr><td>{@code 1 * 2 / 3 / 4}</td><td>{@code ((1 * 2) / 3 / 4)}</td></tr>
  * </table>
  *
+ *
+ * @deprecated Replaced with {@link ASTInfixExpression}
  */
-public final class ASTMultiplicativeExpression extends AbstractLrBinaryExpr {
+@Deprecated
+public final class ASTMultiplicativeExpression extends AbstractJavaExpr {
 
 
     ASTMultiplicativeExpression(int id) {