+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/ASTShiftExpression.java
@@ -28,8 +28,11 @@ package net.sourceforge.pmd.lang.java.ast;
  *     <tr><td>{@code 1 >> 2 << 3 << 4}</td><td>{@code ((1 >> 2) << 3 << 4)}</td></tr>
  *     <tr><td>{@code 1 >> 2 << 3 << 4 >> 5}</td><td>{@code (((1 >> 2) << 3 << 4) >> 5)}</td></tr>
  * </table>
+ *
+ * @deprecated Replaced with {@link ASTInfixExpression}
  */
-public final class ASTShiftExpression extends AbstractLrBinaryExpr implements ASTExpression {
+@Deprecated
+public final class ASTShiftExpression extends AbstractJavaExpr implements ASTExpression {
     ASTShiftExpression(int id) {
         super(id);
     }