+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/AbstractJavaRule.java
@@ -14,19 +14,29 @@ import net.sourceforge.pmd.lang.ast.AstProcessingStage;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.java.JavaLanguageModule;
 import net.sourceforge.pmd.lang.java.JavaProcessingStage;
+import net.sourceforge.pmd.lang.java.ast.ASTAdditiveExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTAllocationExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTAndExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTAnnotation;
 import net.sourceforge.pmd.lang.java.ast.ASTAnnotationMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTArguments;
 import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTCompilationUnit;
+import net.sourceforge.pmd.lang.java.ast.ASTConditionalAndExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTConditionalOrExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTEqualityExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTExclusiveOrExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTInclusiveOrExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTLiteral;
+import net.sourceforge.pmd.lang.java.ast.ASTMultiplicativeExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclarator;
 import net.sourceforge.pmd.lang.java.ast.ASTPrimaryExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTPrimaryPrefix;
 import net.sourceforge.pmd.lang.java.ast.ASTPrimarySuffix;
+import net.sourceforge.pmd.lang.java.ast.ASTRelationalExpression;
+import net.sourceforge.pmd.lang.java.ast.ASTShiftExpression;
 import net.sourceforge.pmd.lang.java.ast.ASTTypeArgument;
 import net.sourceforge.pmd.lang.java.ast.ASTUnaryExpressionNotPlusMinus;
 import net.sourceforge.pmd.lang.java.ast.ASTWildcardBounds;
@@ -98,8 +108,13 @@ public abstract class AbstractJavaRule extends AbstractRule implements JavaParse
         return false;
     }
 
+    /**
+     * @deprecated Not useful, and suppression should happen transparently to rule implementations.
+     *             This will be removed with 7.0.0
+     */
+    @Deprecated
     protected boolean isSuppressed(Node node) {
-        return JavaRuleViolation.isSupressed(node, this);
+        return false;
     }
 
 
@@ -130,9 +145,60 @@ public abstract class AbstractJavaRule extends AbstractRule implements JavaParse
     // REMOVE ME
     // deprecated stuff kept for compatibility with existing visitors, not matched by anything
 
+    @Override
     @Deprecated
     public Object visit(ASTAnnotationMethodDeclaration node, Object data) {
         return null;
+    }
+
+    @Deprecated
+    public Object visit(ASTConditionalOrExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTConditionalAndExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTInclusiveOrExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTExclusiveOrExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTAndExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTEqualityExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTRelationalExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTShiftExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTAdditiveExpression node, Object data) {
+        return visit((ASTExpression) node, data);
+    }
+
+    @Deprecated
+    public Object visit(ASTMultiplicativeExpression node, Object data) {
+        return visit((ASTExpression) node, data);
     }
 
     @Deprecated