+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/rule/codestyle/UnnecessaryFullyQualifiedNameRule.java
@@ -7,13 +7,14 @@ package net.sourceforge.pmd.lang.java.rule.codestyle;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.Set;
 
 import org.apache.commons.lang3.StringUtils;
 
+import net.sourceforge.pmd.RuleContext;
 import net.sourceforge.pmd.lang.java.ast.ASTClassOrInterfaceType;
-import net.sourceforge.pmd.lang.java.ast.ASTCompilationUnit;
 import net.sourceforge.pmd.lang.java.ast.ASTImportDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTName;
 import net.sourceforge.pmd.lang.java.ast.ASTPackageDeclaration;
@@ -24,6 +25,9 @@ import net.sourceforge.pmd.lang.java.ast.JavaNode;
 import net.sourceforge.pmd.lang.java.ast.TypeNode;
 import net.sourceforge.pmd.lang.java.rule.AbstractJavaRule;
 import net.sourceforge.pmd.lang.java.symboltable.SourceFileScope;
+import net.sourceforge.pmd.lang.java.symboltable.VariableNameDeclaration;
+import net.sourceforge.pmd.lang.symboltable.NameOccurrence;
+import net.sourceforge.pmd.lang.symboltable.Scope;
 
 public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
 
@@ -31,7 +35,6 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
     private String currentPackage;
 
     public UnnecessaryFullyQualifiedNameRule() {
-        super.addRuleChainVisit(ASTCompilationUnit.class);
         super.addRuleChainVisit(ASTPackageDeclaration.class);
         super.addRuleChainVisit(ASTImportDeclaration.class);
         super.addRuleChainVisit(ASTClassOrInterfaceType.class);
@@ -39,10 +42,9 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
     }
 
     @Override
-    public Object visit(ASTCompilationUnit node, Object data) {
+    public void start(final RuleContext ctx) {
         imports.clear();
         currentPackage = null;
-        return data;
     }
 
     @Override
@@ -103,7 +105,15 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
     }
 
     private void checkImports(TypeNode node, Object data) {
-        String name = node.getImage();
+        final String name = node.getImage();
+
+        // variable names shadow everything else
+        // If the first segment is a variable, then all
+        // the following are field accesses and it's not an FQCN
+        if (isVariable(node.getScope(), name)) {
+            return;
+        }
+
         List<ASTImportDeclaration> matches = new ArrayList<>();
 
         // Find all "matching" import declarations
@@ -172,7 +182,7 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
         if (matches.isEmpty()) {
             if (isJavaLangImplicit(node)) {
                 addViolation(data, node, new Object[] { node.getImage(), "java.lang.*", "implicit "});
-            } else if (isSamePackage(node)) {
+            } else if (isSamePackage(node, name)) {
                 addViolation(data, node, new Object[] { node.getImage(), currentPackage + ".*", "same package "});
             }
         } else {
@@ -211,11 +221,52 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
         return result;
     }
 
-    private boolean isSamePackage(TypeNode node) {
-        String name = node.getImage();
-        return name.substring(0, name.lastIndexOf('.')).equals(currentPackage);
+    private boolean isVariable(Scope scope, String name) {
+        String firstSegment = name.substring(0, name.indexOf('.'));
+
+        while (scope != null) {
+
+            for (Entry<VariableNameDeclaration, List<NameOccurrence>> entry : scope.getDeclarations(VariableNameDeclaration.class).entrySet()) {
+                if (entry.getKey().getName().equals(firstSegment)) {
+                    return true;
+                }
+            }
+
+            scope = scope.getParent();
+        }
+
+        return false;
     }
-    
+
+    private boolean isSamePackage(TypeNode node, String name) {
+        if (node.getType() != null) {
+            // with type resolution we can do an exact package match
+            Package packageOfType = node.getType().getPackage();
+            if (packageOfType != null) {
+                
+                // get "package" candidate from name
+                int i = name.lastIndexOf('.');
+                if (i > 0) {
+                    name = name.substring(0, i);
+                }
+                
+                return node.getType().getPackage().getName().equals(currentPackage)
+                        && name.equals(currentPackage);
+            }
+        }
+
+        int i = name.lastIndexOf('.');
+        while (i > 0) {
+            name = name.substring(0, i);
+            if (name.equals(currentPackage)) {
+                return true;
+            }
+            i = name.lastIndexOf('.');
+        }
+
+        return false;
+    }
+
     private boolean isJavaLangImplicit(TypeNode node) {
         String name = node.getImage();
         boolean isJavaLang = name != null && name.startsWith("java.lang.");
@@ -233,7 +284,7 @@ public class UnnecessaryFullyQualifiedNameRule extends AbstractJavaRule {
     }
 
     private boolean isAvoidingConflict(final TypeNode node, final String name,
-            final ASTImportDeclaration firstMatch) {
+                                       final ASTImportDeclaration firstMatch) {
         // is it a conflict between different imports?
         if (firstMatch.isImportOnDemand() && firstMatch.isStatic()) {
             final String methodCalled = name.substring(name.indexOf('.') + 1);