+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/ParserCornersTest.java
@@ -20,6 +20,7 @@ import java.util.Map;
 
 import org.apache.commons.io.IOUtils;
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import net.sourceforge.pmd.PMD;
@@ -41,7 +42,7 @@ public class ParserCornersTest {
                 + "                TestInnerClassCallsOuterParent.super.toString();\n" + "            }\n"
                 + "        };\n" + "    }\n" + "}\n");
     }
-    
+
     /**
      * #888 PMD 6.0.0 can't parse valid <> under 1.8.
      */
@@ -106,13 +107,10 @@ public class ParserCornersTest {
         String test18 = readAsString("/net/sourceforge/pmd/ast/ParserCornerCases18.java");
         ASTCompilationUnit cu = parseJava18(test18);
 
-        Assert.assertEquals(13, cu.findChildNodesWithXPath("//FormalParameter").size());
+        Assert.assertEquals(9, cu.findChildNodesWithXPath("//FormalParameter").size());
         Assert.assertEquals(32, cu.findChildNodesWithXPath("//LambdaParameter").size());
         Assert.assertEquals(8, cu.findChildNodesWithXPath("//LambdaParameter[@TypeInferred='false']").size());
-        Assert.assertEquals(4,
-                cu.findChildNodesWithXPath("//FormalParameter[@ExplicitReceiverParameter='true']").size());
-        Assert.assertEquals(9,
-                cu.findChildNodesWithXPath("//FormalParameter[@ExplicitReceiverParameter='false']").size());
+        Assert.assertEquals(4, cu.findChildNodesWithXPath("//ReceiverParameter").size());
     }
 
     @Test
@@ -181,7 +179,7 @@ public class ParserCornersTest {
         String c = readAsString("Bug1530.java");
         parseJava18(c);
     }
-    
+
     @Test
     public void testGitHubBug207() {
         String c = readAsString("GitHubBug207.java");
@@ -202,7 +200,7 @@ public class ParserCornersTest {
         String c = readAsString("GitHubBug208.java");
         parseJava15(c);
     }
-    
+
     @Test
     public void testGitHubBug257NonExistingCast() throws Exception {
         String code = "public class Test {" + PMD.EOL
@@ -227,7 +225,7 @@ public class ParserCornersTest {
     /**
      * This triggered bug #1484 UnusedLocalVariable - false positive -
      * parenthesis
-     * 
+     *
      * @throws Exception
      */
     @Test
@@ -262,6 +260,7 @@ public class ParserCornersTest {
     }
 
     @Test
+    @Ignore("this test depends on usage resolution")
     public void testMethodReferenceConfused() throws Exception {
         String code = readAsString("MethodReferenceConfused.java");
         ASTCompilationUnit compilationUnit = ParserTstUtil.parseAndTypeResolveJava("10", code);