+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/KotlinTestingDsl.kt
@@ -16,10 +16,13 @@ import io.kotlintest.should as kotlintestShould
  * Represents the different Java language versions.
  */
 enum class JavaVersion : Comparable<JavaVersion> {
-    J1_3, J1_4, J1_5, J1_6, J1_7, J1_8, J9, J10, J11, J12;
+    J1_3, J1_4, J1_5, J1_6, J1_7, J1_8, J9, J10, J11, J12, J12__PREVIEW, J13, J13__PREVIEW;
 
     /** Name suitable for use with e.g. [ParserTstUtil.parseAndTypeResolveJava] */
-    val pmdName: String = name.removePrefix("J").replace('_', '.')
+    val pmdName: String = name.removePrefix("J").replaceFirst("__", "-").replace('_', '.').toLowerCase()
+
+
+    operator fun not(): List<JavaVersion> = values().toList() - this
 
     /**
      * Overloads the range operator, e.g. (`J9..J11`).
@@ -352,13 +355,13 @@ open class ParserTestCtx(val javaVersion: JavaVersion = JavaVersion.Latest,
 
             override fun getTemplate(construct: String, ctx: ParserTestCtx): String =
                 """
-                ${ctx.imports.joinToString(separator = "\n")}
-                ${ctx.genClassHeader} {
-                    {
-                        Object o = $construct;
-                    }
-                }
-                """.trimIndent()
+${ctx.imports.joinToString(separator = "\n")}
+${ctx.genClassHeader} {
+    {
+        Object o = $construct;
+    }
+}
+                """
 
 
             override fun retrieveNode(acu: ASTCompilationUnit): ASTExpression = acu.getFirstDescendantOfType(ASTExpression::class.java)!!