+++ docs/pages/release_notes.md
@@ -21,14 +21,24 @@ This is a {{ site.pmd.release_type }} release.
 
 ### Fixed Issues
 
+*   java-codestyle
+    *    [#2017](https://github.com/pmd/pmd/issues/2017): \[java] UnnecessaryFullyQualifiedName triggered for inner class
+
 ### API Changes
+
 
 #### Deprecated APIs
 
 ##### For removal
 
-*   The method {% jdoc java::ast.ASTImportDeclaration#getPackage() %} has been deprecated and
-    will be removed with PMD 7.0.0.
+* pmd-java
+  * {% jdoc java::lang.java.ast.CanSuppressWarnings %} and its implementations
+  * {% jdoc java::lang.java.rule.AbstractJavaRule#isSuppressed(Node) %}
+  * {% jdoc java::lang.java.rule.JavaRuleViolation#isSupressed(Node,Rule) %}
+* pmd-apex
+  * {% jdoc java::lang.apex.ast.CanSuppressWarnings %} and its implementations
+  * {% jdoc java::lang.apex.rule.ApexRuleViolation#isSupressed(Node,Rule) %}
+
 
 ### External Contributions